fix: remove schema default 6000 + dedup formatAge
Review #1 🔴: dcp.schema.json had nudgeGrowthTokens:6000 in both the
field default and the compress default object. If any config layer
applies schema defaults, it would override resolveAdaptiveNudgeGrowth.
Removed from both locations — field has no default, description documents
adaptive behavior.

Review #2 🔴: formatBlockAge was duplicated in status.ts and nudge.ts.
Extracted to lib/ui/utils.ts as formatAge, imported in both modules.
