+++ dcp.schema.json
@@ -278,8 +278,7 @@
                 },
                 "nudgeGrowthTokens": {
                     "type": "number",
-                    "default": 6000,
-                    "description": "Token growth required between per-message nudges (absolute, not percentage). Shows Tips only after context grows by this many tokens. Default is adaptive: 5% of modelContextLimit, clamped to [6000, 50000] — e.g. 128K→6.4K, 200K→10K, 1M→50K."
+                    "description": "Token growth required between per-message nudges (absolute, not percentage). Shows Tips only after context grows by this many tokens. If unset, default is adaptive: 5% of modelContextLimit, clamped to [6000, 50000] — e.g. 128K→6.4K, 200K→10K, 1M→50K. Must be unset (not 6000) for adaptive default to work."
                 }
             },
             "default": {
@@ -296,8 +295,7 @@
                 "protectTags": false,
                 "protectUserMessages": false,
                 "minNudgeContextPercent": 15,
-                "maxSummaryLengthHard": 3000,
-                "nudgeGrowthTokens": 6000
+                "maxSummaryLengthHard": 3000
             }
         },
         "gc": {
