+++ lib/prompts/extensions/nudge.ts
@@ -1,13 +1,6 @@
 import type { SessionState, CompressionBlock } from "../../state"
 import type { GCConfig } from "../../config"
-
-function formatBlockAge(createdAt: number): string {
-    const elapsed = Date.now() - createdAt
-    if (elapsed < 60_000) return "just now"
-    if (elapsed < 3_600_000) return `${Math.floor(elapsed / 60_000)}m ago`
-    if (elapsed < 86_400_000) return `${Math.floor(elapsed / 3_600_000)}h ago`
-    return `${Math.floor(elapsed / 86_400_000)}d ago`
-}
+import { formatAge as formatBlockAge } from "../../ui/utils"
 
 export interface BlockGuidanceContext {
     currentTokens?: number
